nvme_test_generator/requirement.txt
pytest>=7.0.0
pynvme>=2.3.2
python3-dev
python3-pip